+++ app/Http/Controllers/SupplierController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Rekening;
 use App\Models\Supplier;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Redirect;
@@ -10,7 +11,7 @@ use Illuminate\Http\Request;
 class SupplierController extends Controller
 {
     public function index_supplier() {
-        $data_supplier = Supplier::filternama()->paginate(10);
+        $data_supplier = Supplier::filternama()->latest()->paginate(10)->withQueryString();
 
         return \view('supplier.index', [
             'judul_index_supplier' => 'List Data Supplier',
@@ -37,7 +38,9 @@ class SupplierController extends Controller
         $supplier->kontak_supplier = $request->kontak_supplier;
         $supplier->save();
 
-        return Redirect::route('supplier.index');
+        return Redirect::route('supplier.index')
+                        ->with('flash_message', 'Supplier Berhasil Disimpan')
+                        ->with('flash_type', 'Saved!');
     }
 
     public function edit_supplier($id) {
@@ -61,13 +64,21 @@ class SupplierController extends Controller
         $supplier->kontak_supplier = $request->kontak_supplier;
         $supplier->save();
 
-        return Redirect::route('supplier.index');
+        return Redirect::route('supplier.index')
+                        ->with('flash_message', 'Supplier Berhasil Diubah')
+                        ->with('flash_type', 'Updated!');
     }
 
     public function delete_supplier($id) : RedirectResponse {
+        $cek_rekening = Rekening::where('supplier_id', $id)->exists();
+        if ($cek_rekening) {
+            return Redirect::back()->with('delete_error', 'Supplier tidak dapat dihapus karena digunakan dalam Tabel Rekening');
+        }
         $supplier = Supplier::find($id);
         $supplier->delete();
 
-        return Redirect::route('supplier.index');
+        return Redirect::route('supplier.index')
+                        ->with('flash_message', 'Supplier Berhasil Dihapus')
+                        ->with('flash_type', 'Deleted!');
     }
 }
